Wyciagniecie klas z kontrolera

+++ SnakeController/SnakeInterface.hpp
@@ -66,7 +66,7 @@ struct ScoreInd
 {
     static constexpr std::uint32_t MESSAGE_ID = 0x70;
 
-    unsigned score;
+    unsigned long  score;
 };
 
 struct LooseInd
